US1407 Fixed small problem with pagination and added testing around view

+++ src/Web/Sfa.Das.Sas.ApplicationServices/Models/ApprenticeshipSearchResults.cs
@@ -4,7 +4,7 @@ namespace Sfa.Das.Sas.ApplicationServices.Models
 {
     public sealed class ApprenticeshipSearchResults
     {
-        public long TotalResults { get; set; }
+        public double TotalResults { get; set; }
 
         public int ResultsToTake { get; set; }
